Create Trailing Sell Percentage Drop Strategy
Look at this example of trailing Profit change due to % profit gain.

DEFAULT_trailing_profit = 2.25 
If 3% Profit reached the DEFAULT_trailing_profit = 2.0   (drops 0.25 of Default)
If 5% Profit reached the DEFAULT_trailing_profit = 1.75 (drops 0.50 of Default)
If 8% Profit reached the DEFAULT_trailing_profit = 1.25 (drops 1.00 of Default)
If 10% Profit reached the DEFAULT_trailing_profit = .75 (drops 1.25 of Default)
If 15% Profit reached the DEFAULT_trailing_profit = .15 (drops 2.10 of Default) ETC
Setting to personnal prefference....

Using this feature, it would give you a better % profit when you are at a higher gain. 
So when hitting 10% profit and then it drops, you will only drop .75% to make a sell instead of the original 2.25% = saving 1.5% profit.
This would be a better strategy for those who want to make a higher percentage for short and long-term trading. 

Comments and improvements welcome.